+++ src/emu/device.h
@@ -10,8 +10,8 @@ class Device {
  public:
   virtual ~Device(){};
 
-  virtual int64_t GetClockFrequency() = 0;
-  virtual int64_t Execute(int64_t cycles) = 0;
+  virtual uint32_t GetClockFrequency() = 0;
+  virtual uint32_t Execute(uint32_t cycles) = 0;
 };
 }
 }
